Fixed communities page layout after wallet merge
