+++ src/app/(dynamicPages)/profile/[username]/_components/index.ts
@@ -6,6 +6,5 @@ export * from "./profile-entries-list";
 export * from "./curation-trail";
 export * from "./profile-referrals";
 export * from "./profile-settings";
-export * from "./profile-communities";
 export * from "./profile-search";
 export * from "./profile-search-content";
